fix(review-fix-loop): extract shared update_pr/local_commit test fixtures
## Summary
- Add `scripts/tests/helpers.py`: the module loader, a bare local
  repository, the always-passing validation commands, the
  marker-file-driven fake reviewer, and the accepting decider/fixer shared
  between `test_local_commit.py` and `test_update_pr.py`.
- Update both test files to import these fixtures from `helpers.py` instead
  of duplicating them, matching `carve-changesets/scripts/tests/helpers.py`'s
  established precedent for one skill's own shared test fixture module.
- Update `references/update-pr.md` and both test files' module docstrings to
  describe the shared-fixture layout.

## Why
- Closes the one code-simplicity gap (`strong_recommendation`) the first
  review-code-change pass on #100 found: ~150 lines of near-verbatim
  duplicated test setup between the two suites.
