+++ skills/review-fix-loop/scripts/tests/test_local_commit.py
@@ -15,49 +15,44 @@ or more fix cycles, budget exhaustion, validation failure (both the
 "unavailable" and "no tractable correction" shapes), operator input
 (rejected/deferred disposition and scope expansion), and recovery from an
 interrupted attempt.
+
+The module loader, a bare local repository, the always-passing validation
+commands, the marker-file-driven fake reviewer, and the accepting
+decider/fixer are shared with `test_update_pr.py` via this sibling
+directory's own `helpers.py`, matching `carve-changesets/scripts/tests/
+helpers.py`'s established precedent for one skill's own shared test fixture
+module.
 """
 
 from __future__ import annotations
 
-import importlib.util
-import sys
 import tempfile
 import unittest
 from pathlib import Path
 from typing import Any
 
-SKILL_ROOT = Path(__file__).resolve().parents[2]
+import helpers
 
+LC = helpers.load_module("review_fix_loop_local_commit", "local_commit.py")
+LE = helpers.load_module(
+    "review_fix_loop_local_execution_for_tests", "local_execution.py"
+)
+VALIDATE = helpers.load_module("review_fix_loop_validate_for_tests", "validate.py")
 
-def _load(name: str, filename: str):
-    spec = importlib.util.spec_from_file_location(
-        name, SKILL_ROOT / "scripts" / filename
-    )
-    assert spec and spec.loader
-    module = importlib.util.module_from_spec(spec)
-    sys.modules[name] = module
-    spec.loader.exec_module(module)
-    return module
-
-
-LC = _load("review_fix_loop_local_commit", "local_commit.py")
-LE = _load("review_fix_loop_local_execution_for_tests", "local_execution.py")
-VALIDATE = _load("review_fix_loop_validate_for_tests", "validate.py")
+# Fixtures shared with test_update_pr.py; see helpers.py.
+init_repo = helpers.init_repo
+ALWAYS_PASS_VALIDATION = helpers.ALWAYS_PASS_VALIDATION
+CLEAN_TEMPLATE = helpers.CLEAN_TEMPLATE
+FINDING_ID = helpers.FINDING_ID
+_finding = helpers.finding
+make_marker_reviewer = helpers.make_marker_reviewer
+fixing_apply_fix = helpers.fixing_apply_fix
+accepting_decide = helpers.accepting_decide
 
 
 # ---------------------------------------------------------------------------
 # Repository fixtures
 # ---------------------------------------------------------------------------
-
-
-def init_repo(path: Path) -> None:
-    path.mkdir(parents=True, exist_ok=True)
-    LE.git("init", "-q", "-b", "main", cwd=path)
-    LE.git("config", "user.email", "test@example.com", cwd=path)
-    LE.git("config", "user.name", "Test", cwd=path)
-    (path / "README.md").write_text("initial\n")
-    LE.git("add", "-A", cwd=path)
-    LE.git("commit", "-q", "-m", "initial commit", cwd=path)
 
 
 def start_candidate(
@@ -78,11 +73,6 @@ def start_candidate(
     head_sha = LE.current_head(repo)
     return base_sha, head_sha
 
-
-ALWAYS_PASS_VALIDATION = [
-    {"name": "focused unit test", "command": "true", "scope": "focused"},
-    {"name": "full repository gate", "command": "true", "scope": "full"},
-]
 
 FLAG_GATED_VALIDATION = [
     {
@@ -162,106 +152,10 @@ def make_invocation(
     }
 
 
-CLEAN_TEMPLATE = {
-    "schema_version": "1.4",
-    "lens": "aggregate",
-    "verdict": "clean",
-    "findings": [],
-    "blocking_reasons": [],
-    "validation_limitations": [],
-    "next_action": "No changes are required.",
-}
-
-FINDING_ID = "correctness-001"
-
-
-def _finding() -> dict[str, Any]:
-    return {
-        "id": FINDING_ID,
-        "lens": "correctness",
-        "severity": "blocking",
-        "confidence": "high",
-        "rule": "example rule",
-        "evidence": [
-            {"location": "marker.txt:1", "detail": "marker.txt is not 'fixed'"}
-        ],
-        "concern": "marker.txt does not read 'fixed'",
-        "impact": "the candidate is incomplete",
-        "proposed_change": "write 'fixed' into marker.txt",
-        "expected_effect": "marker.txt reads 'fixed'",
-    }
-
-
-def make_marker_reviewer(repo: Path):
-    """A fake reviewer whose verdict is a real function of `marker.txt`'s
-    content at the exact head it is asked to review: `clean` when it reads
-    'fixed', `changes_required` with one `correctness-001` finding otherwise.
-    """
-
-    def reviewer(
-        *, packet, briefing, head_sha, comparison_base_sha, independence, sequence
-    ) -> LC.ReviewPass:
-        del packet, briefing, independence, sequence
-        content = LE.git("show", f"{head_sha}:marker.txt", cwd=repo).stdout.strip()
-        candidate = {"head_sha": head_sha, "comparison_base_sha": comparison_base_sha}
-        if content == "fixed":
-            result = {
-                **CLEAN_TEMPLATE,
-                "candidate": candidate,
-                "lens_executions": [
-                    {
-                        "lens": lens,
-                        "head_sha": head_sha,
-                        "comparison_base_sha": comparison_base_sha,
-                        "verdict": "clean",
-                        "freshly_executed": True,
-                    }
-                    for lens in (
-                        "solution_simplicity",
-                        "correctness",
-                        "code_simplicity",
-                    )
-                ],
-            }
-        else:
-            result = {
-                **CLEAN_TEMPLATE,
-                "candidate": candidate,
-                "verdict": "changes_required",
-                "findings": [_finding()],
-                "lens_executions": [
-                    {
-                        "lens": "solution_simplicity",
-                        "head_sha": head_sha,
-                        "comparison_base_sha": comparison_base_sha,
-                        "verdict": "clean",
-                        "freshly_executed": True,
-                    }
-                ],
-                "next_action": f"Fix {FINDING_ID}.",
-            }
-        return LC.ReviewPass(result=result)
-
-    return reviewer
-
-
-def fixing_apply_fix(*, finding, attempt_path, change_contract, attempt_number):
-    del finding, change_contract, attempt_number
-    (attempt_path / "marker.txt").write_text("fixed\n")
-    return f"fix: resolve {FINDING_ID}"
-
-
 def ineffective_apply_fix(*, finding, attempt_path, change_contract, attempt_number):
     del finding, change_contract
     (attempt_path / "marker.txt").write_text(f"still-broken-{attempt_number}\n")
     return f"fix attempt {attempt_number} for {FINDING_ID}"
-
-
-def accepting_decide(*, finding, change_contract, attempt_number):
-    del change_contract, attempt_number
-    return LC.FixDecision(
-        disposition="accepted", rationale=f"{finding['id']} is tractable"
-    )
 
 
 class LocalCommitRepoTestCase(unittest.TestCase):
