+++ skills/review-fix-loop/scripts/tests/test_update_pr.py
@@ -10,6 +10,15 @@ matching `carve-changesets/scripts/tests/helpers.py`'s established
 state" convention. No test in this file, and no code path `update_pr.py`
 itself can reach, ever touches this repository's real `origin` remote.
 
+Fixtures shared with `test_local_commit.py` (the module loader, a bare local
+repository, the always-passing validation commands, the marker-file-driven
+fake reviewer, and the accepting decider/fixer) live in this sibling
+directory's own `helpers.py` rather than being duplicated here — following
+the same precedent `carve-changesets/scripts/tests/helpers.py` already set
+for that skill. Only the parts genuinely specific to `update_pr` (the
+disposable bare-remote fixture, `start_candidate`'s push, and
+`make_invocation`'s publication/source-binding shape) live in this file.
+
 Covers the ticket's required scenarios: a successful converge-then-publish
 run; a stale expected-old remote state; local non-fast-forward history
 relative to the recorded expected-old head; a misconfigured ("wrong")
@@ -21,46 +30,33 @@ well-formed grant does not itself block a run.
 
 from __future__ import annotations
 
-import importlib.util
-import sys
 import tempfile
 import unittest
 from pathlib import Path
 from typing import Any
 
-SKILL_ROOT = Path(__file__).resolve().parents[2]
+import helpers
 
-
-def _load(name: str, filename: str):
-    spec = importlib.util.spec_from_file_location(
-        name, SKILL_ROOT / "scripts" / filename
-    )
-    assert spec and spec.loader
-    module = importlib.util.module_from_spec(spec)
-    sys.modules[name] = module
-    spec.loader.exec_module(module)
-    return module
-
-
-UP = _load("review_fix_loop_update_pr", "update_pr.py")
+UP = helpers.load_module("review_fix_loop_update_pr", "update_pr.py")
 LC = UP.LC
 LE = UP.LE
 VALIDATE = UP.VALIDATE
+
+# Fixtures shared with test_local_commit.py; see helpers.py.
+init_repo = helpers.init_repo
+ALWAYS_PASS_VALIDATION = helpers.ALWAYS_PASS_VALIDATION
+CLEAN_TEMPLATE = helpers.CLEAN_TEMPLATE
+FINDING_ID = helpers.FINDING_ID
+_finding = helpers.finding
+make_marker_reviewer = helpers.make_marker_reviewer
+make_clean_reviewer = helpers.make_clean_reviewer
+fixing_apply_fix = helpers.fixing_apply_fix
+accepting_decide = helpers.accepting_decide
 
 
 # ---------------------------------------------------------------------------
 # Repository + disposable remote fixtures
 # ---------------------------------------------------------------------------
-
-
-def init_repo(path: Path) -> None:
-    path.mkdir(parents=True, exist_ok=True)
-    LE.git("init", "-q", "-b", "main", cwd=path)
-    LE.git("config", "user.email", "test@example.com", cwd=path)
-    LE.git("config", "user.name", "Test", cwd=path)
-    (path / "README.md").write_text("initial\n")
-    LE.git("add", "-A", cwd=path)
-    LE.git("commit", "-q", "-m", "initial commit", cwd=path)
 
 
 def init_bare_remote(root: Path, *, name: str = "remote.git") -> Path:
@@ -92,12 +88,6 @@ def start_candidate(
     if push_head:
         LE.git("push", str(bare), f"{branch}:refs/heads/{branch}", cwd=repo)
     return base_sha, head_sha
-
-
-ALWAYS_PASS_VALIDATION = [
-    {"name": "focused unit test", "command": "true", "scope": "focused"},
-    {"name": "full repository gate", "command": "true", "scope": "full"},
-]
 
 
 def make_invocation(
@@ -170,126 +160,6 @@ def make_invocation(
         "validation": validation or ALWAYS_PASS_VALIDATION,
         "publication": publication,
     }
-
-
-CLEAN_TEMPLATE = {
-    "schema_version": "1.4",
-    "lens": "aggregate",
-    "verdict": "clean",
-    "findings": [],
-    "blocking_reasons": [],
-    "validation_limitations": [],
-    "next_action": "No changes are required.",
-}
-
-FINDING_ID = "correctness-001"
-
-
-def _finding() -> dict[str, Any]:
-    return {
-        "id": FINDING_ID,
-        "lens": "correctness",
-        "severity": "blocking",
-        "confidence": "high",
-        "rule": "example rule",
-        "evidence": [
-            {"location": "marker.txt:1", "detail": "marker.txt is not 'fixed'"}
-        ],
-        "concern": "marker.txt does not read 'fixed'",
-        "impact": "the candidate is incomplete",
-        "proposed_change": "write 'fixed' into marker.txt",
-        "expected_effect": "marker.txt reads 'fixed'",
-    }
-
-
-def make_marker_reviewer(repo: Path):
-    """A fake reviewer whose verdict is a real function of `marker.txt`'s
-    content at the exact head it is asked to review, mirroring
-    `test_local_commit.py`'s own fixture of the same name."""
-
-    def reviewer(
-        *, packet, briefing, head_sha, comparison_base_sha, independence, sequence
-    ):
-        del packet, briefing, independence, sequence
-        content = LE.git("show", f"{head_sha}:marker.txt", cwd=repo).stdout.strip()
-        candidate = {"head_sha": head_sha, "comparison_base_sha": comparison_base_sha}
-        if content == "fixed":
-            result = {
-                **CLEAN_TEMPLATE,
-                "candidate": candidate,
-                "lens_executions": [
-                    {
-                        "lens": lens,
-                        "head_sha": head_sha,
-                        "comparison_base_sha": comparison_base_sha,
-                        "verdict": "clean",
-                        "freshly_executed": True,
-                    }
-                    for lens in (
-                        "solution_simplicity",
-                        "correctness",
-                        "code_simplicity",
-                    )
-                ],
-            }
-        else:
-            result = {
-                **CLEAN_TEMPLATE,
-                "candidate": candidate,
-                "verdict": "changes_required",
-                "findings": [_finding()],
-                "lens_executions": [
-                    {
-                        "lens": "solution_simplicity",
-                        "head_sha": head_sha,
-                        "comparison_base_sha": comparison_base_sha,
-                        "verdict": "clean",
-                        "freshly_executed": True,
-                    }
-                ],
-                "next_action": f"Fix {FINDING_ID}.",
-            }
-        return LC.ReviewPass(result=result)
-
-    return reviewer
-
-
-def make_clean_reviewer():
-    def reviewer(
-        *, packet, briefing, head_sha, comparison_base_sha, independence, sequence
-    ):
-        del packet, briefing, independence, sequence
-        candidate = {"head_sha": head_sha, "comparison_base_sha": comparison_base_sha}
-        result = {
-            **CLEAN_TEMPLATE,
-            "candidate": candidate,
-            "lens_executions": [
-                {
-                    "lens": lens,
-                    "head_sha": head_sha,
-                    "comparison_base_sha": comparison_base_sha,
-                    "verdict": "clean",
-                    "freshly_executed": True,
-                }
-                for lens in ("solution_simplicity", "correctness", "code_simplicity")
-            ],
-        }
-        return LC.ReviewPass(result=result)
-
-    return reviewer
-
-
-def fixing_apply_fix(*, finding, attempt_path, change_contract, attempt_number):
-    del finding, change_contract, attempt_number
-    (attempt_path / "marker.txt").write_text("fixed\n")
-    return f"fix: resolve {FINDING_ID}"
-
-
-def accepting_decide(*, finding, change_contract, attempt_number):
-    del change_contract, attempt_number
-    return LC.FixDecision(
-        disposition="accepted", rationale=f"{finding['id']} is tractable"
-    )
 
 
 class UpdatePrRepoTestCase(unittest.TestCase):
